requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
psycopg2-binary
pydantic
python-dotenv
gunicorn
